// webapp/src/api/model/nation.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { NationAdvantageTraitsItem } from './nationAdvantageTraitsItem';
import type { NationDisadvantageTraitsItem } from './nationDisadvantageTraitsItem';

export interface Nation {
  advantageTraits: NationAdvantageTraitsItem[];
  description: string;
  disadvantageTraits: NationDisadvantageTraitsItem[];
  id?: string;
  name: string;
}